Refactor: 임시로 테스트 하는 파일 삭제+ 소켓 코드로 대체
