+++ components/channel/PublicRoomCard.tsx
@@ -10,10 +10,14 @@ export default function PublicRoomCard({
   isLocked: boolean;
 }) {
   return (
-    <div className='flex justify-around bg-cyan-100 rounded-full items-center h-12'>
+    <div className='flex justify-around bg-custom2 rounded-full items-center h-12'>
       <span className='w-[200px]'>{channelName}</span>
       <span className='w-10'>({userCount})</span>
-      <Input className='w-[200px]' placeholder='비밀번호' disabled={isLocked} />
+      <Input
+        className='w-[200px]'
+        placeholder='비밀번호'
+        disabled={!isLocked}
+      />
       <Button>참가하기</Button>
     </div>
   );
