--- components/channel2/ChannelHeader.tsx
@@ -1,19 +1,0 @@
-import CreateChannel from './CreateChannel';
-import ManageChannel from './ManageChannel';
-import PublicRoomList from './PublicRoomList';
-
-export default function ChannelHeader({channel_name}: {channel_name: string}) {
-  return (
-    <div className='flex bg-custom2 items-baseline justify-between w-full'>
-      <div>
-        <PublicRoomList />
-      </div>
-      <div>
-        <CreateChannel />
-      </div>
-      <div>
-        <ManageChannel channel_name={channel_name} />
-      </div>
-    </div>
-  );
-}
